refactor(web): extract VfoPanel + SimControls from DesignSession
Part of issue #642 seam 5b-2. Pulls the measurement-freq VFO block
(band dropdown + LCD, Live/Optimize controls, dial + lock) into a
module-level VfoPanel, with the Live/Optimize sub-block as its own
module-level SimControls component per the design comment. JSX bodies
move verbatim.

The Opt* types (OptObjective, OptimizeResult, OptPause) and the
optimize-objective label/order consts move with the block into
VfoPanel.tsx (byte-identical) since VfoPanel is their only render
consumer; DesignSession now imports the types it still needs for its
own state and the /optimize fetch. optMenuOpen becomes SimControls'
own local state — nothing outside that subtree ever read or wrote it.
